criando crud de pessoa

+++ src/main/java/com/estudospedro/sinerji/model/implement/DAO.java
@@ -14,11 +14,11 @@ import java.util.List;
 public interface DAO <T> {
     	List<T> findAll();
 
-	T findById(Integer id);
+	T findById(long id);
 
 	List<T> findByName(String name);
 
 	boolean save(T t);
 
-	boolean delete(Integer id);
+	boolean delete(long id);
 }
